fix: fix loader decorator executed on client
