+++ app/db/models.ts
@@ -1,4 +1,4 @@
-import { client } from "./client";
+import { client } from "./client.server";
 
 export interface UserTable {
   id: number;
